+++ Presentation/Code/CountSort.h
@@ -22,7 +22,7 @@ class CountSort {
 
     int min() const;
     int max() const;
-    unsigned int width() const;
+    long width() const;
     unsigned long size() const;
     bool empty() const;
     void insert(int);
@@ -78,7 +78,7 @@ class CountSort {
     unsigned int *numbers_;
     const int lower_bound_;
     const int upper_bound_;
-    const unsigned int width_;
+    const long width_;
     unsigned long size_ = 0;
     void check_range_(int, int, int) const;
 };
